refactor: Extract ReplyInput as component
